requirements.txt
bluepy
fitbit
python-dotenv
opsgenie_sdk